requirements.txt
alpaca-py
python-dotenv
pandas
numpy
pytz
pyarrow
duckdb
schedule
alpaca-py
duckdb
python-dotenv
pandas